ProtocolBufferSerializer builds again now, but some functionality need to be reimplemented.
