+++ opencog/persist/zmq/atomspace/CMakeLists.txt
@@ -10,8 +10,8 @@ INCLUDE_DIRECTORIES (
 # I'm disabling it until it is debugged and added to the unit tests or removed:
 #
 ADD_LIBRARY (zmqatoms SHARED
-#    ProtocolBufferSerializer
     ZMQMessages.pb
+    ProtocolBufferSerializer
     ZMQClient
 #    ZMQServer
 	ZMQPersistSCM
